Fix gcc warning

+++ generic/tclTest.c
@@ -3929,7 +3929,7 @@ TestregexpObjCmd(
 
 	    if (ii == TCL_INDEX_NONE) {
 		TclRegExpRangeUniChar(regExpr, ii, &start, &end);
-	    } else if (ii > info.nsubs) {
+	    } else if (ii > (size_t)info.nsubs) {
 		start = -1;
 		end = -1;
 	    } else {
@@ -3954,7 +3954,7 @@ TestregexpObjCmd(
 	    if (ii == TCL_INDEX_NONE) {
 		TclRegExpRangeUniChar(regExpr, ii, &start, &end);
 		newPtr = Tcl_GetRange(objPtr, start, end);
-	    } else if (ii > info.nsubs) {
+	    } else if (ii > (size_t)info.nsubs) {
 		newPtr = Tcl_NewObj();
 	    } else {
 		newPtr = Tcl_GetRange(objPtr, info.matches[ii].start,
